Support nullable values

+++ src/NexusMods.Sdk/Tracking/Events.cs
@@ -10,11 +10,11 @@ public static class Events
         EventPropertyDefinition.Create<uint>("file_id"),
         EventPropertyDefinition.Create<uint>("mod_id"),
         EventPropertyDefinition.Create<uint>("game_id"),
-        EventPropertyDefinition.Create<uint>("mod_uid"),
-        EventPropertyDefinition.Create<uint>("file_uid"),
+        EventPropertyDefinition.Create<ulong>("mod_uid"),
+        EventPropertyDefinition.Create<ulong>("file_uid"),
         EventPropertyDefinition.Create<ulong>("file_size"),
         EventPropertyDefinition.Create<long>("duration_ms"),
-        EventPropertyDefinition.Create<uint>("collection_id", isOptional: true),
-        EventPropertyDefinition.Create<uint>("revision_id", isOptional: true),
+        EventPropertyDefinition.Create<ulong>("collection_id", isOptional: true),
+        EventPropertyDefinition.Create<ulong>("revision_id", isOptional: true),
     };
 }
